LPS-76927 Moving CoreUpgradeProcessRegistry to portal-impl because it is not going to be needed for modules
